q3_financial_intelligence_assistant/requirements.txt
# Core RAG and AI Framework
langchain==0.1.12
langchain-openai==0.1.6
langchain-community==0.0.28
openai>=1.24.0,<2.0.0

# Vector Database
pinecone-client==3.0.2
weaviate-client==3.26.4

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.2.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
alembic==1.13.1

# Caching
redis==5.0.1
celery==5.3.4

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
pandas==2.2.0
numpy==1.26.4

# Async and Concurrency
aiofiles==23.2.1
httpx==0.26.0
asyncio-throttle==1.0.2

# Monitoring and Logging
prometheus-client==0.19.0
structlog==24.1.0

# Testing and Load Testing
locust==2.17.0
pytest==8.0.0
pytest-asyncio==0.23.5

# Utilities
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Additional dependencies for production
gunicorn==21.2.0 